requirements.txt
playsound==1.2.2
torch==2.0.1+cu117
numpy==1.24.3
opencv-python==4.6.0
ultralytic==8.0.181
cvzone==1.5.6
supervision==0.15.0rc1
pandas==1.5.3
Pillow==9.4.0